+++ CHANGES.rst
@@ -5,6 +5,9 @@
   (`Issue #88 <https://github.com/sqlalchemy-redshift/sqlalchemy-redshift/pull/88>`_)
 - Increase max_identifier_length to 127 characters
   (`Issue #96 <https://github.com/sqlalchemy-redshift/sqlalchemy-redshift/issues/96>`_)
+- Fix a bug where table names containing a period caused an error on reflection
+  (`Issue #97 <https://github.com/sqlalchemy-redshift/sqlalchemy-redshift/pull/97>`_)
+
 
 0.5.0 (2016-04-21)
 ------------------